% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract1.R
\name{extract1}
\alias{extract1}
\title{Extract a Single Parameter From a Stanfit Object}
\usage{
extract1(object, par, ...)
}
\arguments{
\item{object}{An object of class stanfit.}

\item{par}{Character string providing the parameter (or other quantity) name of interest.}

\item{...}{Additional arguments to be passed to \code{\link[rstan:extract]{rstan::extract()}}.}
}
\value{
When permuted = TRUE (the default), this function returns an array
representing samples for \code{par} with all chains merged together.
}
\description{
Wrapper for \code{\link[rstan:extract]{rstan::extract()}} making it less cumbersome to extract a single parameter.
}
