% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold_partition.R
\name{kfold_partition}
\alias{kfold_partition}
\title{K-Fold Partition}
\usage{
kfold_partition(psm_dat, K, N_random = 1000)
}
\arguments{
\item{psm_dat}{A data frame with a column of class factor named \code{site}.}

\item{K}{A positive integer giving the number of groups.}

\item{N_random}{A natural number giving the number of random permutations to compare.}
}
\value{
A named list with components \describe{
\item{\code{N_group}}{A vector of group sizes.}
\item{group}{A vector giving the group membership of each row in \code{psm_dat}.}
}
}
\description{
Randomly partition sites into groups for cross-validation.
Sites are randomly partitioned into \emph{K} groups that are
roughly similar in size (i.e., number of observations). If multiple permutations
are requested, \code{kfold_partition} returns the one with the smallest range
of group sizes.
}
