% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_lulc_predict.R
\name{sem_lulc_predict}
\alias{sem_lulc_predict}
\title{Posterior Predictive Distribution of Landscape Variables}
\usage{
sem_lulc_predict(fit, data, d, newZ = NULL, linpred = FALSE)
}
\arguments{
\item{fit}{Object of class \code{stanfit} representing a fitted PSM SEM.
Parameters \code{a0}, \code{A}, and \code{phi} must be monitored.}

\item{data}{The data list passed to \code{stan()} to estimate \code{fit}.}

\item{d}{Integer in {1, ..., \emph{D}} indicating which landscape
variable to simulate.}

\item{newZ}{Matrix of dimension \code{N_new x L} giving the \emph{L} factor
scores for each new prediction. If \code{NULL}, the posterior samples of
site-specific factor scores are used.}

\item{linpred}{Logical indicating whether to return the linear predictor (the
default is \code{FALSE}).}
}
\value{
List with elements \describe{ \item{\code{X}}{An \code{iter x
  N_new} matrix containing posterior samples of the predicted landscape variable.}
  \item{\code{g_mu_X}}{An \code{iter x N_new} matrix containing posterior samples of the linear
  predictor evaluated at the specified values of \emph{Z}.} }
}
\description{
Compute the posterior predictive distribution of landscape variables from a
fitted SEM, possibly conditioned on specified values of the latent factor(s).
}
