% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_psm.R
\name{sem_psm}
\alias{sem_psm}
\title{Fit Structural Equation Model of Pre-Spawning Mortality}
\usage{
sem_psm(
  psm,
  X,
  normal_indx,
  gamma_indx,
  L = 1,
  fit = NULL,
  I0_Z = 1,
  I_su = 1,
  I_su_Z = 1,
  I_fa = 1,
  I_fa_Z = 1,
  I_fit = rep(1, nrow(psm)),
  I_lpd = rep(1, nrow(psm)),
  pars = c("a0", "A", "Z", "phi", "g_mu_X", "mu_b0", "b0_Z", "sigma_b0", "b0",
    "mu_b_su", "b_su_Z", "sigma_b_su", "b_su", "mu_b_fa", "b_fa_Z", "sigma_b_fa", "b_fa",
    "sigma_psm", "p_psm", "ll_psm"),
  chains = 3,
  iter = 12000,
  warmup = 2000,
  thin = 5,
  ...
)
}
\arguments{
\item{psm}{A data frame with rows corresponding to observations, including columns named
\describe{
\item{\code{site}}{A factor giving site names.}
\item{\code{ppt_su}}{A numeric variable giving summer precipitation in mm.}
\item{\code{ppt_fa}}{A numeric variable giving fall precipitation in mm.}
\item{\code{n}}{An integer variable giving the number of carcasss sampled.}
\item{\code{n_psm}}{An integer variable giving the number of pre-spawning mortalities
in each sample.}
}}

\item{X}{A matrix with rows corresponding to sites and columns corresponding to 
landscape variables, to be modeled as either normal- or gamma-distributed.}

\item{normal_indx}{An integer vector giving indices of columns of \code{X} to be modeled
as normally distributed.}

\item{gamma_indx}{An integer vector giving indices of columns of \code{X} to be modeled as
gamma-distributed.}

\item{L}{Integer giving the number of latent landscape factors to fit.}

\item{fit}{Stanfit object as returned by a previous call to [SEMPSM::sem_psm()]. If
not NULL, initial values will be drawn randomly from the posterior samples in \code{fit}
using [SEMPSM::stan_init_cv()]; otherwise they will be generated using [SEMPSM::stan_init()].}

\item{I0_Z}{Integer (0/1) indicating whether to include main effects of latent landscape 
factors on pre-spawning mortality risk.}

\item{I_su}{Integer (0/1) indicating whether to include a main effect of summer precipitation
on pre-spawning mortality risk.}

\item{I_su_Z}{Integer (0/1) indicating whether to include interactions between 
summer precipitation and latent landscape factors on pre-spawning mortality risk.}

\item{I_fa}{Integer (0/1) indicating whether to include a main effect of fall precipitation
on pre-spawning mortality risk.}

\item{I_fa_Z}{Integer (0/1) indicating whether to include interactions between 
fall precipitation and latent landscape factors on pre-spawning mortality risk.}

\item{I_fit}{Integer vector (0/1) of length \code{nrow(psm)} indicating whether each
pre-spawning mortality observation should be included in the likelihood. This may be
useful for fitting some observations while simulating from the posterior predictive
distribution for others.}

\item{I_lpd}{Integer vector (0/1) of length \code{nrow(psm)} indicating whether
to evaluate the log posterior predictive density for each observation.}

\item{pars}{A character vector specifying parameters of interest to be saved.}

\item{chains}{A positive integer specifying the number of Markov chains.}

\item{iter}{A positive integer specifying the number of iterations for each chain (including warmup).}

\item{warmup}{A positive integer specifying the number of warmup (aka burnin) iterations per chain.}

\item{thin}{A positive integer specifying the period for saving samples.}

\item{...}{Other arguments to be passed to [rstan::sampling()].}
}
\value{
An object of S4 class [rstan::stanfit].
}
\description{
Fit a hierarchical Bayesian structural equation model to data on pre-spawning mortality
and landscape and climatic variables using Stan.
}
