% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_psm_predict.R
\name{sem_psm_predict}
\alias{sem_psm_predict}
\title{Posterior Predictive Distribution of PSM Risk}
\usage{
sem_psm_predict(
  fit,
  data,
  newsites = NULL,
  newZ = NULL,
  level = "site",
  transform = FALSE,
  gradient = FALSE
)
}
\arguments{
\item{fit}{Object of class \code{stanfit} representing a fitted PSM SEM. Parameters
\code{mu_b0}, \code{sigma_b0}, \code{b0_std}, and \code{b0_Z} must be monitored.}

\item{data}{The data list passed to \code{\link[rstan:sampling]{rstan::sampling()}} to estimate \code{fit},
as returned by \code{\link[SEMPSM:stan_data]{SEMPSM::stan_data()}}.}

\item{newsites}{Optional numeric vector of length \code{N_new} giving the site numbers
(coded as in \code{data}) for which predictions are to be made. If \code{newsites} is
not specified, the default is to use all sites in \code{data}.}

\item{newZ}{Optional matrix of dimension \code{N_new x L} giving the \emph{L} factor scores
for each new prediction. If \code{NULL}, the posterior samples of site-specific factor
scores are used.}

\item{level}{Level of grouping at which to predict. Options are \code{"site"}
(site-specific interannual average, the default) or \code{"year"}
(include year-within-site residual variation).}

\item{transform}{Logical indicating whether to return the linear predictor (\code{FALSE},
the default) or inverse-logit transform it.}

\item{gradient}{Logical indicating whether to compute the gradient of PSM risk with respect to the
latent factor(s).}
}
\value{
List with elements \describe{
\item{\code{est}}{An \code{iter x N_new} matrix containing posterior samples of the predicted probability
(or the linear predictor of logit probability, if \code{transform == FALSE}) of PSM.}
\item{\code{gradient}}{An \code{iter x N_new x L} array whose \code{[,,l]} panel contains posterior
samples of the gradient of PSM with respect to the \code{l}-th factor, evaluated at \code{newZ}.}
}
}
\description{
Compute the posterior predictive distribution of PSM risk from a fitted SEM,
possibly conditioned on specified values of the latent factor(s).
}
\details{
Currently predictions can only be made for sites (with or without PSM data)
used to fit the model, and for sample-average precipitation conditions (i.e., precipitation effects
are set to zero). These restrictions will be lifted in a later version.
}
