% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEMPSM-package.R
\docType{package}
\name{SEMPSM-package}
\alias{SEMPSM-package}
\alias{SEMPSM}
\title{The 'SEMPSM' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2019). RStan: the R interface to Stan. R package version 2.19.2. https://mc-stan.org
}
