% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_data.R
\name{stan_data}
\alias{stan_data}
\title{Assemble Data for Stan}
\usage{
stan_data(
  psm,
  X,
  normal_indx,
  gamma_indx,
  L = 1,
  I0_Z = 1,
  I_su = 1,
  I_su_Z = 1,
  I_fa = 1,
  I_fa_Z = 1,
  I_fit = rep(1, nrow(psm)),
  I_lpd = rep(1, nrow(psm))
)
}
\arguments{
\item{psm}{A data frame with rows corresponding to observations, including columns named
\describe{
\item{\code{site}}{A factor giving site names.}
\item{\code{ppt_su}}{A numeric variable giving summer precipitation in mm.}
\item{\code{ppt_fa}}{A numeric variable giving fall precipitation in mm.}
\item{\code{n}}{An integer variable giving the number of carcasss sampled.}
\item{\code{n_psm}}{An integer variable giving the number of pre-spawning mortalities
in each sample.}
}}

\item{X}{A matrix with rows corresponding to sites and columns corresponding to 
landscape variables, to be modeled as either normal- or gamma-distributed.}

\item{normal_indx}{An integer vector giving indices of columns of \code{X} to be modeled
as normally distributed.}

\item{gamma_indx}{An integer vector giving indices of columns of \code{X} to be modeled as
gamma-distributed.}

\item{L}{Integer giving the number of latent landscape factors to fit.}

\item{I0_Z}{Integer (0/1) indicating whether to include main effects of latent landscape 
factors on pre-spawning mortality risk.}

\item{I_su}{Integer (0/1) indicating whether to include a main effect of summer precipitation
on pre-spawning mortality risk.}

\item{I_su_Z}{Integer (0/1) indicating whether to include interactions between 
summer precipitation and latent landscape factors on pre-spawning mortality risk.}

\item{I_fa}{Integer (0/1) indicating whether to include a main effect of fall precipitation
on pre-spawning mortality risk.}

\item{I_fa_Z}{Integer (0/1) indicating whether to include interactions between 
fall precipitation and latent landscape factors on pre-spawning mortality risk.}

\item{I_fit}{Integer vector (0/1) of length \code{nrow(psm)} indicating whether each
pre-spawning mortality observation should be included in the likelihood. This may be
useful for fitting some observations while simulating from the posterior predictive
distribution for others.}

\item{I_lpd}{Integer vector (0/1) of length \code{nrow(psm)} indicating whether
to evaluate the log posterior predictive density for each observation.}
}
\value{
A named list of data to be passed to [SEMPSM::sem_psm()].
}
\description{
Assemble list of data to be passed to [rstan::stan()].
}
