% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_init.R
\name{stan_init}
\alias{stan_init}
\title{Generate Initial Parameter Values}
\usage{
stan_init(data)
}
\arguments{
\item{data}{A named list of data to be passed to Stan, as returned by
\code{\link[SEMPSM:stan_data]{SEMPSM::stan_data()}}.}
}
\value{
A named list of initial parameter values to be passed to \code{\link[rstan:sampling]{rstan::sampling()}}.
}
\description{
Generate initial parameter values for fitting a structural equation model
of pre-spawning mortality using Stan.
}
