% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_mean.R
\name{stan_mean}
\alias{stan_mean}
\title{Extract Posterior Mean Across All Chains From Stanfit Object}
\usage{
stan_mean(object, pars)
}
\arguments{
\item{object}{An object of class stanfit.}

\item{pars}{Character string providing the parameter (or other quantity) name of interest.}

\item{...}{Additional arguments to be passed to \code{\link[rstan]{get_posterior_mean}}.}
}
\value{
When permuted = TRUE (the default), this function returns an array 
representing samples for \code{par} with all chains merged together.
}
\description{
Wrapper for \code{\link[rstan]{get_posterior_mean}} making it less cumbersome to extract 
the mean of a parameter (or other quantity) across all chains.
}
