% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vioplot2.R
\name{vioplot2}
\alias{vioplot2}
\title{Modified Violin Plot}
\usage{
vioplot2(
  x,
  ...,
  range = 1.5,
  h = NULL,
  quantiles = NULL,
  ylim = NULL,
  names = NULL,
  horizontal = FALSE,
  col = "lightblue",
  border = "black",
  lty = 1,
  lwd = 1,
  lty.quantile = 1,
  lwd.quantile = 1,
  rectCol = "black",
  colMed = "white",
  pchMed = 19,
  at,
  add = FALSE,
  wex = 1,
  drawRect = TRUE
)
}
\description{
Modified version of \code{\link[vioplot:vioplot]{vioplot::vioplot()}} that plots lines at specified quantiles. See
\code{\link[vioplot:vioplot]{vioplot::vioplot()}} for description of arguments.
}
