% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_init_cv.R
\name{stan_init_cv}
\alias{stan_init_cv}
\title{Generate Initial Parameter Values Using a Fitted Model}
\usage{
stan_init_cv(fit)
}
\arguments{
\item{fit}{A Stanfit object containing posterior samples from a pres-spawning mortality
structural equation model, as returned by [SEMPSM::SEMPSM]. Initial values are drawn
randomly from the joint posterior in \code{fit}.}
}
\value{
A named list of initial parameter values to be passed to[rstan::stan].
}
\description{
Generate Initial Parameter Values Using a Fitted Model
}
